Add watcher known-garbage RTA test in debug_tools suite
